+++ routes/study/search.js
@@ -1,11 +1,11 @@
-const express = require('express');
+const express = require("express");
 const router = express.Router();
-const bodyParser = require('body-parser');
-const db = require('../util/db');
+const bodyParser = require("body-parser");
+const db = require("../../util/db");
 
 router.use(bodyParser.urlencoded({ extended: true }));
 
-router.get('/', (req, res) => {
+router.get("/", (req, res) => {
   const { keyword, sort, page, perPage } = req.query; //페이징처리
   const offset = (page - 1) * perPage;
 
@@ -19,7 +19,7 @@ router.get('/', (req, res) => {
     query += ` WHERE s.title LIKE '%${keyword}%' OR s.content LIKE '%${keyword}%' `;
   }
 
-  if (sort === 'latest') {
+  if (sort === "latest") {
     query += ` ORDER BY s.created_at DESC `;
   } else {
     query += ` ORDER BY s.created_at ASC `;
@@ -29,14 +29,12 @@ router.get('/', (req, res) => {
 
   db.query(query, (error, results) => {
     if (error) {
-      console.error('검색 중 오류 발생:', error);
-      res.status(500).json({ error: '검색 중 오류가 발생했습니다.' });
+      console.error("검색 중 오류 발생:", error);
+      res.status(500).json({ error: "검색 중 오류가 발생했습니다." });
     } else {
       res.json(results);
     }
   });
 });
-
-
 
 module.exports = router;
