+++ routes/study/tag.js
@@ -1,12 +1,12 @@
-const express = require('express');
+const express = require("express");
 const router = express.Router();
-const db = require('../util/db');
+const db = require("../../util/db");
 
 router.use(express.urlencoded({ extended: true }));
 
 // 태그 검색 요청 처리
-router.get('/', (req, res) => {
-  console.log('GET /tag route called');
+router.get("/", (req, res) => {
+  console.log("GET /tag route called");
   // const searchTag = req.query.tag;
   const searchTag = "백엔드";
   console.log(`검색된 태그: ${searchTag}`);
@@ -23,14 +23,13 @@ router.get('/', (req, res) => {
 
   db.query(query, [decodedSearchTag], (error, results) => {
     if (error) {
-      console.error('태그 검색 중 오류 발생:', error);
-      res.status(500).json({ error: '태그 검색 중 오류가 발생했습니다.' });
+      console.error("태그 검색 중 오류 발생:", error);
+      res.status(500).json({ error: "태그 검색 중 오류가 발생했습니다." });
     } else {
-      console.log('검색 결과:', results);
+      console.log("검색 결과:", results);
       res.json(results);
     }
   });
 });
-
 
 module.exports = router;
